obu: Use uint64_t for OBU and TU sizes.
Compliance with the AV1 spec requires that decoders
support OBU and TU lengths up to 2^56 - 1. Remove usage of
size_t to avoid problems on 32-bit targets, and to ensure
builds for those targets can comply with the spec.

+++ examples/decode_with_drops.c
@@ -112,7 +112,7 @@ int main(int argc, char **argv) {
   while (aom_video_reader_read_frame(reader)) {
     aom_codec_iter_t iter = NULL;
     aom_image_t *img = NULL;
-    size_t frame_size = 0;
+    uint64_t frame_size = 0;
     int skip;
     const unsigned char *frame =
         aom_video_reader_get_frame(reader, &frame_size);
