+++ ivfdec.h
@@ -19,8 +19,8 @@ extern "C" {
 
 int file_is_ivf(struct AvxInputContext *input);
 
-int ivf_read_frame(FILE *infile, uint8_t **buffer, size_t *bytes_read,
-                   size_t *buffer_size);
+int ivf_read_frame(FILE *infile, uint8_t **buffer, uint64_t *bytes_read,
+                   uint64_t *buffer_size);
 
 #ifdef __cplusplus
 } /* extern "C" */
