+++ webmdec.h
@@ -55,7 +55,7 @@ int file_is_webm(struct WebmInputContext *webm_ctx,
 //      1 - End of Stream
 //     -1 - Error
 int webm_read_frame(struct WebmInputContext *webm_ctx, uint8_t **buffer,
-                    size_t *bytes_read, size_t *buffer_size);
+                    uint64_t *bytes_read, uint64_t *buffer_size);
 
 // Guesses the frame rate of the input file based on the container timestamps.
 int webm_guess_framerate(struct WebmInputContext *webm_ctx,
